Problem solved. In accepted use fragment you could start one work, and end another. but now user can stop only work that he started

+++ app/src/main/java/com/example/user/jobche/ApiRequests.kt
@@ -66,7 +66,7 @@ interface ApiRequests {
     @GET("tasks/{taskId}/applications")
     fun getAppliers(
         @Header("Authorization") auth: String,
-        @Path("taskId") taskId: Int,
+        @Path("taskId") taskId: Long,
         @Query("page") page: Int,
         @Query("size") size: Int
     ): Call<Applications>
